+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Popover adjusted to positioning on the page.
+
 ## [1.1.0] - 2025-05-16
 
 ### Added